requirements.txt
asgiref==3.8.1
certifi==2024.12.14
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
Django==5.1.5
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docker==7.1.0
dockerize==0.2.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pluggy==1.5.0
psycopg==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyelftools==0.31
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
